promptpilot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
groq==0.4.2 